+++ src/main/java/web/services/ScreenshotService.java
@@ -27,6 +27,9 @@ public class ScreenshotService {
     private FilmService filmService;
 
     public ScreenshotDb getScreenshotWithId(String id) throws HibernateException, IndexOutOfBoundsException {
+        if (id == null || id.isEmpty()) {
+            throw new IllegalArgumentException("Id should not be null or empty");
+        }
         Session session = sessionFactory.getCurrentSession();
         ScreenshotDb screenshotDb =
                 (ScreenshotDb) session.createQuery("from ScreenshotDb s where s.id=" + id).list().get(0);
@@ -34,11 +37,17 @@ public class ScreenshotService {
     }
 
     public void saveOrUpdate(ScreenshotDb screenshotDb) throws HibernateException {
+        if (screenshotDb == null) {
+            throw new IllegalArgumentException("ScreenshotDb should not be null");
+        }
         Session session = sessionFactory.getCurrentSession();
         session.saveOrUpdate(screenshotDb);
     }
 
     public void delete(String id) throws HibernateException {
+        if (id == null || id.isEmpty()) {
+            throw new IllegalArgumentException("Id should not be null or empty");
+        }
         Session session = sessionFactory.getCurrentSession();
         session.createQuery("delete from ScreenshotDb s where s.id=" + id).executeUpdate();
     }
@@ -50,6 +59,9 @@ public class ScreenshotService {
     }
 
     public Set<ScreenshotDb> createScreenshotSet(List<String> links) throws ParsingJsonToDaoException {
+        if (links == null) {
+            throw new IllegalArgumentException("Links should not be null");
+        }
         Set<ScreenshotDb> screenshotDbSet = new HashSet<>();
         for (String s : links) {
             ScreenshotDb screenshotDb = new ScreenshotDb();
@@ -60,6 +72,12 @@ public class ScreenshotService {
     }
 
     public void checkForSceens(int idFilm, Set<ScreenshotDb> screenshotDbSet) throws HibernateException {
+        if (idFilm < 0) {
+            throw new IllegalArgumentException("IdFilm should not be null");
+        }
+        if (screenshotDbSet == null) {
+            throw new IllegalArgumentException("ScreenshotDbSet should not be null");
+        }
         Session session = sessionFactory.getCurrentSession();
         session.createQuery("delete from ScreenshotDb s where s.filmByIdFilm=" + idFilm).executeUpdate();
         for (ScreenshotDb screenshotDb : screenshotDbSet) {
