+++ src/main/java/web/services/TrailerService.java
@@ -28,6 +28,9 @@ public class TrailerService {
     private FilmService filmService;
 
     public TrailerDb getTrailerWithId(String id) throws HibernateException, IndexOutOfBoundsException {
+        if (id == null || id.isEmpty()) {
+            throw new IllegalArgumentException("Id should not be null or empty");
+        }
         Session session = sessionFactory.getCurrentSession();
         TrailerDb trailerDb =
                 (TrailerDb) session.createQuery("from TrailerDb t where t.id=" + id).list().get(0);
@@ -35,11 +38,17 @@ public class TrailerService {
     }
 
     public void saveOrUpdate(TrailerDb trailerDb) throws HibernateException {
+        if (trailerDb == null) {
+            throw new IllegalArgumentException("TrailerDb should not be null");
+        }
         Session session = sessionFactory.getCurrentSession();
         session.saveOrUpdate(trailerDb);
     }
 
     public void delete(String id) throws HibernateException {
+        if (id == null || id.isEmpty()) {
+            throw new IllegalArgumentException("Id should not be null or empty");
+        }
         Session session = sessionFactory.getCurrentSession();
         session.createQuery("delete from TrailerDb t where t.id=" + id).executeUpdate();
     }
@@ -51,6 +60,9 @@ public class TrailerService {
     }
 
     public Set<TrailerDb> createTrailerDbSet(List<String> trailers) throws ParsingJsonToDaoException {
+        if (trailers == null) {
+            throw new IllegalArgumentException("Trailers should not be null");
+        }
         Set<TrailerDb> trailerDbSet = new HashSet<>();
         for (String s : trailers) {
             TrailerDb trailerDb = new TrailerDb();
@@ -61,6 +73,12 @@ public class TrailerService {
     }
 
     public void checkForTrailers(int idFilm, Set<TrailerDb> trailerDbSet) throws HibernateException {
+        if (idFilm < 0) {
+            throw new IllegalArgumentException("IdFilm should not be null");
+        }
+        if (trailerDbSet == null) {
+            throw new IllegalArgumentException("TrailerDbSer should not be null");
+        }
         Session session = sessionFactory.getCurrentSession();
         session.createQuery("delete from TrailerDb t where t.filmByIdFilm=" + idFilm).executeUpdate();
         for (TrailerDb trailerDb : trailerDbSet) {
